--- a/sunpy/net/cdaweb/helpers.py
+++ b/sunpy/net/cdaweb/helpers.py
@@ -2,6 +2,7 @@ import json
 import pathlib
 
 import requests
+from concurrent.futures import ThreadPoolExecutor
 
 from astropy.table import Table
 
@@ -125,13 +126,23 @@ def _update_cdaweb_dataset_data():
     ])
     # Mapping from dataset ID to description
     all_datasets = {}
-    # Generate list of tuples with query parameters
-    query_parameters = [('observatoryGroup', group) for group in all_obs['Group']]
-    # Send request and fetch data
-    resp = requests.get(url, params=query_parameters, headers=_CDAS_HEADERS)
-    datasets = resp.json()['DatasetDescription']
-    dataset_ids = {ds['Id']: ds['Label'] for ds in datasets}
-    all_datasets.update(dataset_ids)
+    # Number of parallel threads we spawn
+    N = 3
+    def _fetch_cdaweb_dataset(group, url=url):
+        print(f'🛰 Getting datasets for {group}')
+        u = url + f'?observatoryGroup={group}'
+        res = requests.get(u, headers=_CDAS_HEADERS)
+        datasets = res.json()['DatasetDescription']
+        dataset_ids = {ds['Id']: ds['Label'] for ds in datasets}
+        all_datasets.update(dataset_ids)
+
+    with ThreadPoolExecutor(max_workers=N) as executor:
+        # Submit each URL to the thread pool
+        futures = [executor.submit(_fetch_cdaweb_dataset, group) for group in all_obs['Group']]
+        # Wait for all tasks to complete
+        for future in futures:
+            future.result()
+
     attr_file = pathlib.Path(__file__).parent / 'data' / 'attrs.json'
     with open(attr_file, 'w') as attrs_file:
         json.dump(dict(sorted(all_datasets.items())), attrs_file, indent=2)