--- a/tools/update_attrs_json.py
+++ b/tools/update_attrs_json.py
@@ -7,10 +7,8 @@ from sunpy.net.vso import VSOClient
 
 print(f"Updating the attrs json files using sunpy {sunpy.__version__}...")
 
-
 print("Updating VSO json...")
 VSOClient.create_parse_vso_values()
-
 
 print("Updating JSOC json...\nThis may take some time...")
 JSOCClient.create_parse_jsoc_values()