Reported in issue #6519 fetching data from cdaweb is slow because we are sending one request per instrument. Whit this patch, code will build query containing all instruments and send request.

There is support for that kind of queries in cdas docs where it says that if you add multiple query parameters with same key, it will OR them https://cdaweb.gsfc.nasa.gov/WebServices/REST/#Get_Datasets

Also, i've ditched the idea of running multiple requests because there is a note here https://cdaweb.gsfc.nasa.gov/WebServices/REST/py/cdasws/ saying that calls can be throttled if we have more than 5 calls simultaneously.

Simple test with `time` and running command with my connections gives around 3 minutes faster code execution

```
# Before patch, one request per instrument
python tools/update_attrs_json.py  8.37s user 0.86s system 1% cpu 8:56.05 total

# After patch, only one request
python tools/update_attrs_json.py  2.08s user 0.75s system 0% cpu 6:09.15 total
```