@lzivadinovic 2022-11-15
> Thanks a lot for this! Could you remove the changelog entry? (this isn't a user-facing tool) Otherwise I will give this a test later this week, but looks +1

Sure, no problem.