@dstansby 2022-11-16
I've just run this, and it might have sped it up a bit, but not significantly. There's also a lack of progress printed now - in the previous version each observatory was printed to the command line when it was being updated, which provided useful feedback that something was happening. Unless we can get this script down to running in <~ 30secs, I think having the feedback is important to reassure that something is happening.

So I think I'm actually 👎 on this (sorry!), unless there's a way to get continual feedback printed during the request being made.