@lzivadinovic 2022-11-16
I dont think there is a way to get progress, this is one api request and we wait while backend process it. We dont get any info, simple get and that's it.

We can try threads but we should be limited to 3 or 4 at given time. 

With gnu parallel and 4 connections i can get
```
time parallel -j4 'curl -H "Accept: application/json" {}' < plist >> res  
# plist is list of endpoints with different observatories
12.12s user 3.74s system 7% cpu 3:40.27 total
```
And with single request: 
```
curl -H "Accept: application/json" https....... -o /dev/null  0.33s user 0.25s system 0% cpu 6:27.84 total
```
So yeah, looks like we can do it this way. I can try this in few days.