@dstansby 2023-01-04
@lzivadinovic do you think it's best to close this PR, and you can always open a new PR with the other approach? Or are you planning on modifying this PR to use the other approach?