@lzivadinovic 2023-03-02
Hi, sorry for very late response. I've added threadpool executor that spawns multiple threads and make multiple parallel request to rest API. It's 'user-friendly' in terms that it will write data its currently fetching, like original solution. Number of connection is limited to N=3 I've read this number somewhere in documentation, that they tolerate that much simultaneous requests from same ip, but this number can be easily changed.